system/jarvis-voice/requirements.txt
# JARVIS Voice System - Dependencies
# ==================================
# pip install -r requirements.txt

# Core
python-dotenv>=1.0.0

# Web Server (Command Center)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# Speech-to-Text (Deepgram)
websockets>=12.0
aiohttp>=3.9.0

# Text-to-Speech (ElevenLabs - fallback)
elevenlabs>=1.0.0
requests>=2.31.0

# Text-to-Speech (Sesame CSM-1B - ultra-humanizado)
# Requer GPU NVIDIA com CUDA para melhor performance
transformers>=4.52.1
torch>=2.0.0
torchaudio>=2.0.0
scipy>=1.10.0

# Claude API
anthropic>=0.18.0

# Audio Processing
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# Optional: For better audio support
# PyAudio>=0.2.14  # Uncomment if sounddevice doesn't work
